ref: combine fst code into single file
